requirements.txt
fpdf
python-docx
pypdf2
boto3
streamlit
pymupdf